+++ lib/ui/src/components/preview/preview.tsx
@@ -1,32 +1,31 @@
-import window from 'global';
-import React, { Component, Fragment, FunctionComponent, ReactNode } from 'react';
+import React, { Component, Fragment, FunctionComponent } from 'react';
 import memoize from 'memoizerific';
-import copy from 'copy-to-clipboard';
 import { styled } from '@storybook/theming';
-import { Consumer, State, API, Combo } from '@storybook/api';
+import { Consumer, API, Combo } from '@storybook/api';
 import { SET_CURRENT_STORY } from '@storybook/core-events';
 import addons, { types, Types, Addon } from '@storybook/addons';
 import merge from '@storybook/api/dist/lib/merge';
-import { Icons, IconButton, Loader, TabButton, TabBar, Separator } from '@storybook/components';
+import { Loader } from '@storybook/components';
 
 import { Helmet } from 'react-helmet-async';
 
-import { ViewMode } from '@storybook/api/dist/modules/addons';
 import { Toolbar } from './toolbar';
 
 import * as S from './components';
 
-import { ZoomProvider, ZoomConsumer, Zoom } from './zoom';
+import { ZoomProvider, ZoomConsumer } from './zoom';
 
 import { IFrame } from './iframe';
+import { PreviewProps, ActualPreviewProps, Wrapper, IframeRenderer } from './PreviewProps';
+import { getTools } from './getTools';
 
-const DesktopOnly = styled.span({
+export const DesktopOnly = styled.span({
   // Hides full screen icon at mobile breakpoint defined in app.js
   '@media (max-width: 599px)': {
     display: 'none',
   },
 });
-const stringifyQueryParams = (queryParams: Record<string, string>) =>
+export const stringifyQueryParams = (queryParams: Record<string, string>) =>
   Object.entries(queryParams).reduce((acc, [k, v]) => {
     return `${acc}&${k}=${v}`;
   }, '');
@@ -42,41 +41,11 @@ const renderIframe: IframeRenderer = (storyId, viewMode, id, baseUrl, scale, que
   />
 );
 
-type IframeRenderer = (
-  storyId: string,
-  viewMode: State['viewMode'],
-  id: string,
-  baseUrl: string,
-  scale: number,
-  queryParams: Record<string, any>
-) => ReactNode;
-
-const getElementList = memoize(
+export const getElementList = memoize(
   10
 )((getFn: API['getElements'], type: Types, base: Partial<Addon>[]) =>
   base.concat(Object.values(getFn(type)))
 );
-
-interface WrapperProps {
-  index: number;
-  children: ReactNode;
-  id: string;
-  storyId: string;
-  active: boolean;
-}
-type Wrapper = { render: FunctionComponent<WrapperProps> };
-
-interface ActualPreviewProps {
-  wrappers: Wrapper[];
-  viewMode: State['viewMode'];
-  id: string;
-  storyId: string;
-  active: boolean;
-  baseUrl: string;
-  scale: number;
-  queryParams: Record<string, any>;
-  customCanvas?: IframeRenderer;
-}
 
 const ActualPreview: FunctionComponent<ActualPreviewProps> = ({
   wrappers,
@@ -123,140 +92,6 @@ const defaultWrappers = [
   } as Wrapper,
 ];
 
-const getTools = memoize(10)(
-  (
-    getElements: API['getElements'],
-    queryParams: State['customQueryParams'],
-    panels: Partial<Addon>[],
-    api: API,
-    options,
-    storyId: string,
-    viewMode: State['viewMode'],
-    docsOnly: boolean,
-    location: State['location'],
-    path: string,
-    baseUrl: string
-  ) => {
-    const tools = getElementList(getElements, types.TOOL, [
-      panels.filter(p => !p.hidden).length > 1
-        ? ({
-            render: () => (
-              <Fragment>
-                <TabBar key="tabs">
-                  {panels
-                    .filter(p => !p.hidden)
-                    .map((t, index) => {
-                      const to = t.route({ storyId, viewMode, path, location });
-                      const isActive = path === to;
-                      return (
-                        <S.UnstyledLink key={t.id || `l${index}`} to={to}>
-                          <TabButton disabled={t.disabled} active={isActive}>
-                            {t.title}
-                          </TabButton>
-                        </S.UnstyledLink>
-                      );
-                    })}
-                </TabBar>
-                <Separator />
-              </Fragment>
-            ),
-          } as Partial<Addon>)
-        : null,
-      {
-        match: p => p.viewMode === 'story',
-        render: () => (
-          <Fragment>
-            <ZoomConsumer>
-              {({ set, value }) => (
-                <Zoom key="zoom" set={(v: number) => set(value * v)} reset={() => set(1)} />
-              )}
-            </ZoomConsumer>
-            <Separator />
-          </Fragment>
-        ),
-      },
-    ]);
-
-    const extraTools = getElementList(getElements, types.TOOLEXTRA, [
-      {
-        match: p => p.viewMode === 'story',
-        render: () => (
-          <DesktopOnly>
-            <IconButton
-              key="full"
-              onClick={api.toggleFullscreen as any}
-              title={options.isFullscreen ? 'Exit full screen' : 'Go full screen'}
-            >
-              <Icons icon={options.isFullscreen ? 'close' : 'expand'} />
-            </IconButton>
-          </DesktopOnly>
-        ),
-      },
-      {
-        match: p => p.viewMode === 'story',
-        render: () => (
-          <IconButton
-            key="opener"
-            href={`${baseUrl}?id=${storyId}${stringifyQueryParams(queryParams)}`}
-            target="_blank"
-            title="Open canvas in new tab"
-          >
-            <Icons icon="share" />
-          </IconButton>
-        ),
-      },
-      {
-        match: p => p.viewMode === 'story',
-        render: () => (
-          <IconButton
-            key="copy"
-            onClick={() =>
-              copy(
-                `${window.location.origin}${
-                  window.location.pathname
-                }${baseUrl}?id=${storyId}${stringifyQueryParams(queryParams)}`
-              )
-            }
-            title="Copy canvas link"
-          >
-            <Icons icon="copy" />
-          </IconButton>
-        ),
-      },
-    ]);
-    // if its a docsOnly page, even the 'story' view mode is considered 'docs'
-    const filter = (item: Partial<Addon>) =>
-      item &&
-      (!item.match ||
-        item.match({
-          storyId,
-          viewMode: docsOnly && viewMode === 'story' ? 'docs' : viewMode,
-          location,
-          path,
-        }));
-
-    const displayItems = (list: Partial<Addon>[]) =>
-      list.reduce(
-        (acc, item, index) =>
-          item ? (
-            // @ts-ignore
-            <Fragment key={item.id || item.key || `f-${index}`}>
-              {acc}
-              {item.render({}) || item}
-            </Fragment>
-          ) : (
-            acc
-          ),
-        null
-      );
-
-    const left = displayItems(tools.filter(filter));
-    const right = displayItems(extraTools.filter(filter));
-
-    return { left, right };
-  }
-);
-
 const getDocumentTitle = (description: string) => {
   return description ? `${description} ⋅ Storybook` : 'Storybook';
 };
@@ -264,27 +99,6 @@ const getDocumentTitle = (description: string) => {
 const mapper = ({ state }: Combo) => ({
   loading: !state.storiesConfigured,
 });
-
-export interface PreviewProps {
-  api: API;
-  storyId: string;
-  viewMode: ViewMode;
-  docsOnly: boolean;
-  options: {
-    isFullscreen: boolean;
-    isToolshown: boolean;
-  };
-  id: string;
-  path: string;
-  location: State['location'];
-  queryParams: State['customQueryParams'];
-  getElements: API['getElements'];
-  customCanvas?: IframeRenderer;
-  description: string;
-  baseUrl: string;
-  parameters: Record<string, any>;
-  withLoader: boolean;
-}
 
 class Preview extends Component<PreviewProps> {
   shouldComponentUpdate({
