+++ lib/ui/src/containers/preview.tsx
@@ -4,7 +4,8 @@ import React from 'react';
 import { Consumer, Combo } from '@storybook/api';
 
 import { StoriesHash } from '@storybook/api/dist/modules/stories';
-import { Preview, PreviewProps } from '../components/preview/preview';
+import { Preview } from '../components/preview/preview';
+import { PreviewProps } from '../components/preview/PreviewProps';
 
 const nonAlphanumSpace = /[^a-z0-9 ]/gi;
 const doubleSpace = /\s\s/gi;
